Move volume data fetching data to volumeGet method
