+++ packages/admin-ui/server-mock/methods/index.ts
@@ -3,3 +3,4 @@ export * from "./devices";
 export * from "./packages";
 export * from "./pending";
 export * from "./userActionLogs";
+export * from "./volumes";
