+++ packages/admin-ui/server-mock/methods/pending.ts
@@ -9,10 +9,9 @@ import {
   NewFeatureStatus,
   PackageNotificationDb,
   PackageNotification,
-  SystemInfo,
-  VolumeData
+  SystemInfo
 } from "../../src/common";
-import { pause } from "../utils";
+import { mountpoints } from "../mockData";
 
 /**
  * Generates a backup of a package and sends it to the client for download.
@@ -155,7 +154,7 @@ export async function getStats(): Promise<HostStats> {
  * by running a pre-written script in the host
  */
 export async function mountpointsGet(): Promise<MountpointData[]> {
-  return [];
+  return mountpoints;
 }
 
 /**
@@ -317,62 +316,4 @@ export async function systemInfoGet(): Promise<SystemInfo> {
       // "change-host-password"
     ]
   };
-}
-
-/**
- * Removes a docker volume by name
- * @param name Full volume name: "bitcoindnpdappnodeeth_bitcoin_data"
- */
-export async function volumeRemove(kwargs: { name: string }): Promise<void> {
-  await pause(1000);
-}
-
-/**
- * Returns volume data
- */
-export async function volumesGet(): Promise<VolumeData[]> {
-  return [
-    {
-      name: "gethdnpdappnodeeth_data",
-      owner: undefined,
-      nameDisplay: "data",
-      ownerDisplay: "gethdnpdappnodeeth",
-      createdAt: 1569346006000,
-      mountpoint: "",
-      size: 161254123,
-      refCount: 0,
-      isOrphan: true
-    },
-    {
-      name: "lightning-networkpublicdappnodeeth_data",
-      owner: "lightning-network.public.dappnode.eth",
-      nameDisplay: "data",
-      ownerDisplay: "lightning-networkpublicdappnodeeth",
-      createdAt: 1569146006000,
-      mountpoint: "/media/usb0",
-      size: 0,
-      fileSystem: {
-        mountpoint: "/media/usb0",
-        use: "89%",
-        used: 198642520,
-        total: 235782040,
-        free: 25092776,
-        vendor: "SanDisk",
-        model: "Ultra_USB_3.0"
-      },
-      refCount: 2,
-      isOrphan: false
-    },
-    {
-      name: "d19f0771fe2e5b813cf0d138a77eddc33ae3fd6afc1cc6daf0fba42ed73e36ae",
-      owner: undefined,
-      nameDisplay: "",
-      ownerDisplay: "",
-      createdAt: 1569306006000,
-      mountpoint: "",
-      size: 24,
-      refCount: 0,
-      isOrphan: true
-    }
-  ];
 }
